Solve merge conflicts and comments of pull request
