+++ tests/unit/__tests__/components/Column.test.tsx
@@ -5,7 +5,7 @@ import { Column, ColumnProps } from '../../../../components/Column';
 import { projectData } from '../../../../utils/interfaces';
 // eslint-disable-next-line import/extensions
 import 'jest-styled-components';
-import { createProjectData } from '../../../../utils/fileSystemUtilities';
+import { createProjectData } from '../../../../utils/testUtilities';
 
 expect.extend(toHaveNoViolations);
